requirements.txt
 
python-owasp-zap-v2.4
